version-finale

+++ medi-doc/src/views/AdminView.vue
@@ -395,6 +395,10 @@
 
       async deletePatient(patientID) {
   try {
+    // Confirmation alert
+    const confirmed = confirm("Are you sure you want to delete the patient?");
+    if (!confirmed) return; // If not confirmed, exit the function
+
     // Delete patient from Firestore
     await deleteDoc(doc(a, patientID));
